+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- Added the Dutch linguistic acceptability dataset `dutch-cola`. It has been set to
+  `unofficial` for now, but it might eventually replace ScaLA-nl as the official Dutch
+  linguistic acceptability dataset. For now, you can benchmark models on it by
+  explicitly setting the dataset using the `--dataset` argument (or `dataset` argument
+  if using the `Benchmarker` API).
 - Added the Belebele dataset, being a multilingual multiple-choice reading comprehension
   dataset. This has been added as a separate `multiple-choice-reading-comprehension`
   task, and is available in all supported languages except Faroese. The dataset has been